// src/types/equipment.types.ts
export enum StanceType {
  Defensive = 0,
  Balanced = 1,
  Offensive = 2,
}

export enum WeaponType {
  ArmingSwordKite = 0,
  MaceTower = 1,
  RapierBuckler = 2,
  Greatsword = 3,
  Battleaxe = 4,
  Quarterstaff = 5,
  Spear = 6,
  ShortswordBuckler = 7,
  ShortswordTower = 8,
  Daggers = 9,
  RapierDagger = 10,
  ScimitarBuckler = 11,
  AxeKite = 12,
  AxeTower = 13,
  Scimitars = 14,
  FlailBuckler = 15,
  MaceKite = 16,
  ClubTower = 17,
  Clubs = 18,
  ArmingSwordShortsword = 19,
  ScimitarDagger = 20,
  ArmingSwordClub = 21,
  AxeMace = 22,
  FlailDagger = 23,
  MaceShortsword = 24,
  Maul = 25,
  Trident = 26,
}

export enum ArmorType {
  Cloth = 0,
  Leather = 1,
  Chain = 2,
  Plate = 3,
}